Use a new-style AddToPlaylistAction in room history
