+++ src/components/RoomHistory/HistoryActions.js
@@ -1,29 +1,15 @@
 import cx from 'clsx';
 import React from 'react';
 import PropTypes from 'prop-types';
-import { useDispatch, useSelector } from 'react-redux';
-import { addMediaMenu } from '../../actions/PlaylistActionCreators';
+import { useSelector } from 'react-redux';
 import { isLoggedInSelector } from '../../selectors/userSelectors';
-import { useMediaListContext } from '../MediaList/BaseMediaList';
-import AddToPlaylistAction from '../MediaList/Actions/AddToPlaylist';
 import PreviewMediaAction from '../MediaList/PreviewMediaAction';
-
-const {
-  useCallback,
-} = React;
+import AddToPlaylistAction from './AddToPlaylistAction';
 
 const dontBubble = (event) => event.stopPropagation();
 
 function HistoryActions({ className, historyEntry }) {
-  const { selection } = useMediaListContext();
   const isLoggedIn = useSelector(isLoggedInSelector);
-
-  const dispatch = useDispatch();
-  const handleAdd = useCallback((position) => {
-    const selectedItems = selection.isSelected(historyEntry) ? selection.get() : [historyEntry];
-
-    dispatch(addMediaMenu(selectedItems.map((entry) => entry.media), position));
-  }, [dispatch, selection, historyEntry]);
 
   if (!isLoggedIn) {
     return null;
@@ -37,7 +23,7 @@ function HistoryActions({ className, historyEntry }) {
       onClick={dontBubble}
     >
       <PreviewMediaAction media={historyEntry.media} />
-      <AddToPlaylistAction onAdd={handleAdd} />
+      <AddToPlaylistAction historyEntry={historyEntry} />
     </div>
   );
 }
